# openapi.yaml
openapi: 3.1.0
info:
  title: DesignInterview Backend API
  version: 1.0.0
  description: |
    HTTP API expected by the frontend API facade in
    `frontend/src/lib/mock-backend.ts`. Realtime document, presence, session,
    and timer messages are transported over the session WebSocket and are
    therefore outside OpenAPI's HTTP operation model.
servers:
  - url: /api
...
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Logged in; the server also establishes the session cookie.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      security: [{ cookieAuth: [] }]
      responses:
        '204': { description: Logged out; no response body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: designinterview_session
      description: Authenticated interviewer session established by login.
    participantAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque participant token
      description: Session-scoped credential established when a guest joins an invitation.
...
  responses:
    BadRequest:
      description: Invalid request.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Authentication is required or invalid.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Conflict:
      description: State or document-version conflict.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    SignupRequest:
      type: object
      additionalProperties: false
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        displayName: { type: string, minLength: 2, maxLength: 60 }
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
...